Fix int64 issues

+++ megatron/core/inference/contexts/attention_context/triton/tensor_ops.py
@@ -109,13 +109,13 @@ def _tensor_masked_update_kernel_2d(
     # 2D Calculation: base + batch * stride0 + col * stride1
     dst_ptr = (
         STATES_PTR
-        + (target_idx.to(tl.int64) * stride_state_b.to(tl.int64))
-        + (row_offsets.to(tl.int64) * stride_state_d0.to(tl.int64))
+        + (target_idx.to(tl.int64) * stride_state_b)
+        + (row_offsets.to(tl.int64) * stride_state_d0)
     )
     src_ptr = (
         NEW_STATES_PTR
         + (pid_batch * stride_new_b.to(tl.int64))
-        + (row_offsets.to(tl.int64) * stride_new_d0.to(tl.int64))
+        + (row_offsets.to(tl.int64) * stride_new_d0)
     )
 
     val = tl.load(src_ptr, mask=mask)
@@ -161,14 +161,14 @@ def _tensor_masked_update_kernel_3d(
     # Calculate pointers using specific strides
     dst_offset = (
         (target_idx.to(tl.int64) * stride_state_b.to(tl.int64))
-        + (idx_d0 * stride_state_d0.to(tl.int64))
-        + (idx_d1 * stride_state_d1.to(tl.int64))
+        + (idx_d0 * stride_state_d0)
+        + (idx_d1 * stride_state_d1)
     )
 
     src_offset = (
         (pid_batch * stride_new_b.to(tl.int64))
-        + (idx_d0 * stride_new_d0.to(tl.int64))
-        + (idx_d1 * stride_new_d1.to(tl.int64))
+        + (idx_d0 * stride_new_d0)
+        + (idx_d1 * stride_new_d1)
     )
 
     dst_ptr = STATES_PTR + dst_offset
@@ -225,16 +225,16 @@ def _tensor_masked_update_kernel_4d(
     # Calculate pointers using specific strides
     dst_offset = (
         (target_idx.to(tl.int64) * stride_state_b.to(tl.int64))
-        + (idx_d0 * stride_state_d0.to(tl.int64))
-        + (idx_d1 * stride_state_d1.to(tl.int64))
-        + (idx_d2 * stride_state_d2.to(tl.int64))
+        + (idx_d0 * stride_state_d0)
+        + (idx_d1 * stride_state_d1)
+        + (idx_d2 * stride_state_d2)
     )
 
     src_offset = (
         (pid_batch * stride_new_b.to(tl.int64))
-        + (idx_d0 * stride_new_d0.to(tl.int64))
-        + (idx_d1 * stride_new_d1.to(tl.int64))
-        + (idx_d2 * stride_new_d2.to(tl.int64))
+        + (idx_d0 * stride_new_d0)
+        + (idx_d1 * stride_new_d1)
+        + (idx_d2 * stride_new_d2)
     )
 
     dst_ptr = STATES_PTR + dst_offset
